requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
pydantic-settings==2.8.1
yfinance==1.2.0
pandas==2.2.3
numpy==2.0.2
prometheus-client==0.22.1
psycopg2-binary==2.9.11
python-dotenv>=1.0.0
redis==5.2.1
httpx==0.28.1
eval_type_backport==0.2.2
tigeropen>=3.5.7
flask>=3.0